+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Deprecated
+
+  * `Entry` fields `key` and `value` are no longer public. Use `entry.key()`
+    and `entry.get()`, respectively, instead.
+
 ## 0.1.0 - 2019-11-13
 
   * Initial release